Create DataModel, move save/load logic there
