+++ Checklists/DataModel.swift
@@ -1,0 +1,54 @@
+//
+//  DataModel.swift
+//  Checklists
+//
+//  Created by Subvert on 20.02.2021.
+//
+
+import Foundation
+
+class DataModel {
+    
+    var lists = [Checklist]()
+    
+    init() {
+        loadChecklists()
+    }
+    
+    // MARK: - Data Saving
+    
+    func documentHistory() -> URL {
+        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
+        return paths[0]
+    }
+    
+    func dataFilePath() -> URL {
+        return documentHistory().appendingPathComponent("Checklists.plist")
+    }
+    
+    // --- SAVING ---
+    func saveChecklists() {
+        let encoder = PropertyListEncoder()
+        do {
+            let data = try encoder.encode(lists)
+            try data.write(to: dataFilePath(), options: Data.WritingOptions.atomic)
+        }
+        catch {
+            print("Error encoding item array: \(error.localizedDescription)")
+        }
+    }
+    
+    // --- LOADING ---
+    func loadChecklists() {
+        let path = dataFilePath()
+        if let data = try? Data(contentsOf: path) {
+            let decoder = PropertyListDecoder()
+            do {
+                lists = try decoder.decode([Checklist].self, from: data)
+            }
+            catch {
+                print("Error decoding item array: \(error.localizedDescription)")
+            }
+        }
+    }
+}
